// src/app/models/clientDetails.model.ts
export interface ClientDetails {
    id: number;
    clientCode: string;
    clientDetailsCode: string;
    createdAt: string;
    entryDate: string;
    leftArm: string;
    rightArm: string;
    chest: string;
    waist: string;
    hips: string;
    leftThigh: string;
    rightThigh: string;
    leftCalf: string;
    rightCalf: string;
    height: any;
    weight: any;
    bfp: any;
    notes: string;
    status: string;
}